[#32927473] Add xPrecision, yPrecision to line-construction.
https://www.pivotaltracker.com/story/show/32927473

It might be that these details should be moved elsewhere, but for now
they live in the line-construction definition.

On Mon, Jul 23, 2012 at 4:42 PM, Anand Balaji <anand.balaji@zeuslearning.com> wrote:

With regard to the new functionality requested within user-story 32927473,
we think that it would be best for the runtime to receive two additional
values from the authoring tool.

Proposed parameters -
1. xPrecision
2. yPrecision

The general thought is that it will be best to separate the plotting/drawing
precision values from the tolerance limits.

Story was:
The tolerance specified in authoring for the precision of the
answer should also control the precision of where the point 'snaps to'
on the graph and the number of decimal places in the point placement
tooltip and the data table.