[#37246813] fix regression: can't copy activities.
From https://www.pivotaltracker.com/story/show/37246813
cstaudt: "I clicked on "Copy this" for Fence Painting and it said "the page you requested could not be found".  Have we lost that function?"

The problem was that pages weren't validating because the predefined graph pane weren't validating (because the point_type was set to 'disc' instead of 'dot'…

The migration which made the switch should have also updated existing records, but it didn't.