Use generator still emitting datadefs with labels,
Labels were removed from datadefs in the (not sure why) in the subsequent commit. However, we want to migrate production to use the multi-dataset version of the generator, and we want to have no diffs in the runtime json emitted by the new and old generator. Therefore, do not use a generator version that changes the runtime output by removing labels from datadefs in the runtime output.