Makes GraphLabel parent polymorphic to enforce single-parent
Following up on this idea: http://stackoverflow.com/a/18726662/306084 which came from the attempt to make a GraphLabel belong to a LabelSet or a PredefinedGraphPane, but not both. Previously we simply didn't offer Labels which already belonged to a LabelSet as options to link to PredefinedGraphPanes, and we'll keep that in place, but now if you do manage to link a GraphLabel to a PredefinedGraphPane, it will be removed from its LabelSet when that happens.