Explicitly set list position of hints
[#60530978]

Before this commit, the list position of multiple choice choices were getting set to incorrect values when deserialized from the has (position values were 25, 26, 27... instead of 1, 2, 3...) I assume this was because associations were not fully set up by the time the position attribute was set during demarshaling.

Tests don't catch this problem. I suspect this is because there is only one activity in the test database at the time activity copying is tested.